Rename otherModes to other-modes
This is the last file that is camelCase.
