+++ source/views/transportation/index.js
@@ -9,7 +9,7 @@ import React from 'react'
 import {TabNavigator} from '../components/tabbed-view'
 import {TabBarIcon} from '../components/tabbar-icon'
 
-import OtherModesView from './otherModes'
+import OtherModesView from './other-modes'
 import BusView from './bus'
 
 export default TabNavigator(
